hotfix: Add a README target to fix autoreconf error
This is a quick fix to the previous commit.  I considered amending
it but it has been published.

Adding a README target in Makefile.am takes care of the GNU coding
standard which requires README (not README.md) to exist - otherwise
autoreconf will fail.  There are other workarounds, but this is one way.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>